Update to latest facebook client and work around an M_UNKNOWN error.
When calling the room creation api, the matrix HS will complain if
the room name specified is set to null, so we pass an empty string instead.

As for why we're getting a null room name from the facebook client api,
I'm not sure. This could be totally normal or there could be a deeper
problem, but I don't have time to look into that at the moment.