Add support for Login Approvals (2FA).
This commit adds support for Login approvals. A login.js script was
added to prompt for username, password, and any relevant login
approval information, log into facebook, and save out the auth token.
Running login.js is now required to set up the bridge, and it completely
replaces storage of any fb username/password info the config.json.

See the README.md for information on how to use it.