Fix a bug with attachment messages that also have a body.
If an fb message had any body text along with one or more attachments,
the attachments would be completely ignored and never sent into the matrix
room -- only the body text would be sent.

This commit fixes that issue, and both the body and any attachments will now
be sent to the matrix room.

This fixes another potential issue where only one attachment would be relayed
to the matrix room if the fb message contains multiple attachments. I was not
able to test this fix, however, because i am not entirely sure how to attach
multiple things to a facebook message.

I should also mention that each fb attachment is sent as a separate message to matrix,
and the body text itself (if any) is also sent as a separate matrix to matrix. I believe
this mirrors how facebook handles things; It doesn't attempt to consolidate everything
into a single message.